requirements.txt
attrs==25.3.0
bandit==1.8.3
blinker==1.9.0
Brotli==1.1.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
coverage==7.8.0
fakeredis==2.28.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.33.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
numpy==2.2.4
openmeteo_requests==1.4.0
openmeteo_sdk==1.19.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pbr==6.1.1
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
psutil==7.0.0
pyee==12.1.1
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-mock==3.14.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rich==14.0.0
robotframework==7.2.2
selenium==4.31.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
stevedore==5.4.1
text-unidecode==1.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
url-normalize==2.2.0
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
